requirements.txt
numpy>=1.23.5
pandas>=1.5.3
nibabel>=5.1.0
scikit-learn>=1.2.2
KDEpy>=1.1.5
dtw-python>=1.3.0
